AU-1910: Split static component logic to base class
